move module hdfmapper.py from `bapsflib.lapd._hdf` to `bapsflib._hdf.maps`, rename module `hdfmapper.py` -> `hdfmap.py`, update 'docs/src' documentation files
