+++ bapsflib/_hdf/maps/hdfmap.py
