+++ docs/hdfmap/main.rst
@@ -1,23 +1,23 @@
 .. _file_mappings:
 
-HDF5 File Mapping (:class:`~bapsflib.lapdhdf.hdfmapper.hdfMap`)
-===============================================================
+HDF5 File Mapping (:class:`~bapsflib._hdf.maps.hdfmap.hdfMap`)
+==============================================================
 
 .. contents:: Contents
     :depth: 3
     :local:
 
-:class:`~bapsflib.lapdhdf.hdfmapper.hdfMap` constructs the mapping for
+:class:`~bapsflib._hdf.map.hdfmap.hdfMap` constructs the mapping for
 a given HDF5 file.  When a HDF5 file is opened with
 :class:`~bapsflib.lapdhdf.files.File`,
-:class:`~bapsflib.lapdhdf.hdfmapper.hdfMap` is automatically called to
+:class:`~bapsflib._hdf.maps.hdfmap.hdfMap` is automatically called to
 construct the map and an instance of the mapping object is bound
 to the file object as :attr:`~bapsflib.lapdhdf.files.File.file_map`.
 Thus, the file mappings for :file:`test.hdf5` can be accessed like::
 
     >>> f = lapdhdf.File('test.hdf5')
     >>> f.file_map
-    <bapsflib.lapdhdf.hdfmapper.hdfMap>
+    <bapsflib._hdf.maps.hdfmap.hdfMap>
 
 .. include:: mapping.inc.rst
 .. include:: add_map_module.inc.rst
