+++ docs/src/bapsflib.lapd._hdf.rst
@@ -10,7 +10,6 @@ bapsflib\.lapd\.\_hdf
 
     bapsflib.lapd._hdf.files
     bapsflib.lapd._hdf.hdferrors
-    bapsflib.lapd._hdf.hdfmapper
     bapsflib.lapd._hdf.hdfoverview
     bapsflib.lapd._hdf.hdfreadcontrol
     bapsflib.lapd._hdf.hdfreaddata
